move preview features panel related components under their own folder
